fix: CalorieRing type mismatch - Float vs Int in remaining calculation

+++ feature-dashboard/src/main/java/com/example/calview/feature/dashboard/components/CalorieRing.kt
@@ -20,8 +20,8 @@ fun CalorieRing(
     goal: Float,
     modifier: Modifier = Modifier
 ) {
-    val progress = (consumed.toFloat() / goal.toFloat()).coerceIn(0f, 1f)
-    val remaining = (goal - consumed).coerceAtLeast(0)
+    val progress = (consumed / goal).coerceIn(0f, 1f)
+    val remaining = (goal - consumed).coerceAtLeast(0f).toInt()
 
     Box(
         modifier = modifier,
